Replace maxAmountOfFiles with maxDepth, as it is a more expected result
When we read files by their number, we never get the right result.
For example, we have the following file system structure:

folder1/
- subfolder1/
--- file1
--- file2
- file3
- file4
folder2/
- subfolder2/
--- file5
--- file6
- file5

and maxAmountOfFiles = 3, then we will read [folder1, subfolder1,
file3, file3, folder2, subfolder2, file5]. And this is not expected
result. It is better to read by depth, which will give the expected
result.